+++ src/main/java/com/algaworks/algafood/domain/service/email/EmailSenderService.java
@@ -2,6 +2,8 @@ package com.algaworks.algafood.domain.service.email;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NonNull;
+import lombok.Singular;
 
 import java.util.Set;
 
@@ -13,8 +15,13 @@ public interface EmailSenderService {
     @Builder
     class Message {
 
+        @Singular
         private Set<String> recipients;
+
+        @NonNull
         private String subject;
+
+        @NonNull
         private String body;
     }
 }
