+++ src/main/resources/application.properties
@@ -37,4 +37,4 @@ spring.mail.host=smtp.sendgrid.net
 spring.mail.port=587
 spring.mail.username=apikey
 
-algafood.email.sender=AlgaFood <noreply@algafood.com>
+algafood.email.sender=AlgaFood <gabrielguarido.oliveira@gmail.com>
